+++ test/format.test.js
@@ -3,6 +3,11 @@ const expect = require('chai').expect
 const Skematic = require('../index')
 const format = require('../src/format')
 
+const gen = {
+  xx: function () { return 'wow' },
+  dbl: function (x) { return x * 2 }
+}
+
 describe('.format(skm, opts, data)', function () {
   // Load in the library of functions
   before(function () {
@@ -27,20 +32,20 @@ describe('.format(skm, opts, data)', function () {
   it('can format scalar values', function () {
     var s = {default: 'smoo'}
 
-    var out = format(s, {}, '')
+    var out = format(s, '')
     expect(out).to.equal('smoo')
   })
 
   it('`sparse` opt defaults to false', function () {
     var s = {name: {default: 'ace'}, power: {default: 3}}
-    var out = format(s, {}, {})
+    var out = format(s, {})
     expect(out.name).to.equal('ace')
     expect(out.power).to.equal(3)
   })
 
   it('`sparse` opt "true" only parses provided data', function () {
     var s = {name: {default: 'ace'}, power: {default: 3}}
-    var out = format(s, {sparse: true}, {power: undefined})
+    var out = format(s, {power: undefined}, {sparse: true})
     expect(out.power).to.equal(3)
     expect(out).to.have.keys('power')
   })
@@ -52,7 +57,7 @@ describe('.format(skm, opts, data)', function () {
   })
 
   it('applies falsey defaults', function () {
-    expect(format({default: false}, undefined)).to.equal(false)
+    expect(format({field: {default: false}}, {field: undefined}).field).to.equal(false)
   })
 
   it('`transform` opt default transforms values', function () {
@@ -63,7 +68,7 @@ describe('.format(skm, opts, data)', function () {
 
   it('`transform` opt "false" does not apply transforms', function () {
     var s = {name: {default: 'ace', transforms: ['uppercase']}}
-    var out = format(s, {transform: false}, {})
+    var out = format(s, {}, {transform: false})
     expect(out.name).to.equal('ace')
   })
 
@@ -71,7 +76,7 @@ describe('.format(skm, opts, data)', function () {
     var s = {
       name: {
         generate: {
-          ops: [{fn: 'xx'}]
+          ops: [{fn: gen.xx}]
         }
       }
     }
@@ -84,48 +89,36 @@ describe('.format(skm, opts, data)', function () {
     var s = {
       name: {
         generate: {
-          ops: [{fn: 'xx'}],
+          ops: [{fn: gen.xx}],
           // NOTE THIS: we're setting name to only generate with "once" flag
           once: true
         }
       }
     }
 
-    var out = format(s, {}, {})
+    var out = format(s, {})
     expect(out).to.not.have.key('name')
 
-    out = format(s, {generate: 'once'}, {})
+    out = format(s, {}, {once: true})
     expect(out.name).to.equal('wow')
   })
 
   it('`generate` opt "false" does not run generators', function () {
     var s = {
       name: {
-        generate: { ops: [{fn: 'xx'}] }
+        generate: { ops: [{fn: gen.xx}] }
       }
     }
 
-    var out = format(s, {generate: false}, {})
+    var out = format(s, {}, {generate: false})
     expect(out).to.not.have.key('name')
   })
 
   it('`strip` removes matching field values', function () {
     var data = {a: undefined, b: null, c: 2, d: ':)'}
     var s = {a: {type: 'number'}}
-    var out = format(s, {strip: [undefined, null, ':)']}, data)
+    var out = format(s, data, {strip: [undefined, null, ':)']})
     expect(out).to.eql({c: 2})
-  })
-
-  it('`copy` clones the data to return', function () {
-    var data = {a: undefined}
-    var out = format({a: {default: 5}}, {copy: true}, data)
-    expect(data.a).to.equal(undefined)
-    expect(out.a).to.equal(5)
-
-    data = [undefined]
-    out = format({type: 'array', schema: {default: 4}}, {copy: true}, data)
-    expect(data[0]).to.equal(undefined)
-    expect(out[0]).to.equal(4)
   })
 
   it('`mapIdFrom` maps the primaryKey from the id field', function () {
@@ -136,7 +129,7 @@ describe('.format(skm, opts, data)', function () {
 
     var data = [ {_id: '512314', name: 'power'}, {_id: '519910', name: 'speed'} ]
 
-    var out = Skematic.format(propSchema, {mapIdFrom: '_id'}, data)
+    var out = Skematic.format(propSchema, data, {mapIdFrom: '_id'})
     expect(out).to.have.length(2)
     expect(out[0]).to.include.key('prop_id')
     expect(out[0]).to.not.include.key('_id')
@@ -149,7 +142,7 @@ describe('.format(skm, opts, data)', function () {
     }
     var data = {woot: '1', name: 'yo', tags: [{whatever: 1, label: 'moo'}]}
 
-    var out = Skematic.format(scm, {strict: true}, data)
+    var out = Skematic.format(scm, data, {strict: true})
     expect(out).to.have.keys('name', 'tags')
     expect(out.tags[0]).to.have.key('label')
   })
@@ -194,7 +187,7 @@ describe('.format(skm, opts, data)', function () {
         person: {
           schema: {
             name: {default: 'Zim'},
-            age: {generate: {ops: [{fn: 'dbl', args: [5]}]}},
+            age: {generate: {ops: [{fn: gen.dbl, args: [5]}]}},
             phrase: {transforms: ['uppercase']}
           },
           // If no default specified, 'person' will only be applied
@@ -207,7 +200,7 @@ describe('.format(skm, opts, data)', function () {
       expect(out.person.name).to.equal('Zim')
       expect(out.person.age).to.equal(10)
 
-      out = format(s, {}, {person: {phrase: 'woo'}})
+      out = format(s, {person: {phrase: 'woo'}})
       expect(out.person.phrase).to.equal('WOO')
     })
   })
@@ -215,21 +208,23 @@ describe('.format(skm, opts, data)', function () {
   describe('arrays', function () {
     it('can array', function () {
       var s = {
-        type: 'array',
-        schema: {
-          default: 'moo',
-          generate: {ops: [{fn: 'xx'}]},
-          transforms: ['uppercase']
-        },
-        default: []
+        field: {
+          type: 'array',
+          schema: {
+            default: 'moo',
+            generate: {ops: [{fn: gen.xx}]},
+            transforms: ['uppercase']
+          },
+          default: []
+        }
       }
 
-      var out = format(s, {}, [undefined, 'hi'])
-      expect(out).to.have.length(2)
-      expect(out[0]).to.equal('WOW')
+      var out = format(s, {field: [undefined, 'hi']})
+      expect(out.field).to.have.length(2)
+      expect(out.field[0]).to.equal('WOW')
 
-      out = format(s, {}, undefined)
-      expect(out).to.eql([])
+      out = format(s, undefined)
+      expect(out.field).to.eql([])
     })
 
     it('support object field values as arrays', function () {
@@ -238,7 +233,7 @@ describe('.format(skm, opts, data)', function () {
           type: 'array',
           schema: {
             default: 'moo',
-            generate: {ops: [{fn: 'xx'}]},
+            generate: {ops: [{fn: gen.xx}]},
             transforms: ['uppercase']
           },
           default: []
@@ -246,11 +241,11 @@ describe('.format(skm, opts, data)', function () {
         moo: {default: []}
       }
 
-      var out = format(s, {}, {zeep: [undefined], moo: [undefined]})
+      var out = format(s, {zeep: [undefined], moo: [undefined]})
       expect(out.zeep[0]).to.equal('WOW')
       expect(out.moo).to.eql([undefined])
 
-      out = format(s, {}, {zeep: undefined})
+      out = format(s, {zeep: undefined})
       expect(out.zeep).to.eql([])
       expect(out.moo).to.eql([])
     })