+++ HISTORY.md
@@ -10,9 +10,15 @@ Added:
 
 Changed:
 
+- Argument order for API methods changed to `(schema, data [, opts])`
+- #25 `.format()` enforces immutability, returns fresh copy of data
 - Rule.required ONLY checks null/undefined
 - Rules 'present', 'notPresent' renamed to 'oneOf', 'notOneOf'
 - Typecast to Date returns ISO 8601 Date '2000-01-01T00:00:00.000Z'
+
+Removed:
+
+- `copy` format option (immutable by default now)
 
 Internal: